+++ packages/agent-infra/mcp-servers/browser/src/tools/tabs.ts
@@ -12,10 +12,12 @@ const newTabTool = defineTool({
     },
   },
   handle: async (ctx, args) => {
-    const { browser } = ctx;
+    const { browser, logger } = ctx;
     try {
       const newPage = await browser!.newPage();
-      await newPage.goto(args.url);
+      await newPage.goto(args.url, {
+        waitUntil: [],
+      });
       await newPage.bringToFront();
 
       // update global browser and page
@@ -28,6 +30,7 @@ const newTabTool = defineTool({
         isError: false,
       };
     } catch (error) {
+      logger.error('Failed to open new tab:', error);
       return {
         content: [
           {
@@ -45,22 +48,43 @@ const tabListTool = defineTool({
   name: 'browser_tab_list',
   config: {
     description: 'Get the list of tabs',
+    outputSchema: {
+      tabList: z.array(
+        z.object({
+          index: z.number(),
+          active: z.boolean(),
+          title: z.string(),
+          url: z.string(),
+        }),
+      ),
+    },
   },
   handle: async (ctx) => {
-    const { browser, page: activePage, currTabsIdx: activePageId } = ctx;
+    const {
+      browser,
+      page: activePage,
+      currTabsIdx: activePageId,
+      logger,
+    } = ctx;
     try {
-      const tabListList = await getTabList(browser);
+      const tabListList = await getTabList(browser, activePageId);
       const tabListSummary =
         tabListList?.length > 0
           ? `Current Tab: [${activePageId}] ${await activePage?.title()}\nAll Tabs: \n${tabListList
-              .map((tab) => `[${tab.index}] ${tab.title} (${tab.url})`)
+              .map(
+                (tab) => `[${tab.index}] Title: ${tab.title} (URL: ${tab.url})`,
+              )
               .join('\n')}`
           : '';
       return {
         content: [{ type: 'text', text: tabListSummary }],
         isError: false,
+        structuredContent: {
+          tabList: tabListList,
+        },
       };
     } catch (error) {
+      logger.error('Failed to browser_tab_list:', error);
       return {
         content: [
           {
@@ -68,6 +92,9 @@ const tabListTool = defineTool({
             text: `Failed to get tab list`,
           },
         ],
+        structuredContent: {
+          tabList: [],
+        },
       };
     }
   },
@@ -82,13 +109,13 @@ const switchTabTool = defineTool({
     },
   },
   handle: async (ctx, args) => {
-    const { browser } = ctx;
+    const { browser, currTabsIdx: activePageId, logger } = ctx;
     try {
       const pages = await browser!.pages();
       if (args.index >= 0 && args.index < pages.length) {
         await pages[args.index].bringToFront();
 
-        const tabListList = await getTabList(browser);
+        const tabListList = await getTabList(browser, activePageId);
         const tabListSummary =
           tabListList?.length > 0
             ? `All Tabs: \n${tabListList
@@ -111,6 +138,7 @@ const switchTabTool = defineTool({
         isError: true,
       };
     } catch (error) {
+      logger.error('Failed to browser_switch_tab:', error);
       return {
         content: [
           {
@@ -130,7 +158,7 @@ const closeTabTool = defineTool({
     description: 'Close the current tab',
   },
   handle: async (ctx) => {
-    const { page, currTabsIdx } = ctx;
+    const { page, currTabsIdx, logger } = ctx;
 
     try {
       await page.close();
@@ -147,6 +175,7 @@ const closeTabTool = defineTool({
         isError: false,
       };
     } catch (error) {
+      logger.error(`Failed to browser_close_tab: [${currTabsIdx}]`, error);
       return {
         content: [
           {