+++ packages/agent-infra/mcp-servers/browser/tests/tools/content.test.ts
@@ -12,6 +12,7 @@ import { InMemoryTransport } from '@modelcontextprotocol/sdk/inMemory.js';
 import { createServer, type GlobalConfig } from '../../src/server.js';
 import express from 'express';
 import { AddressInfo } from 'net';
+import { McpError } from '@modelcontextprotocol/sdk/types.js';
 
 describe('Browser Content Tests', () => {
   let client: Client;
@@ -65,13 +66,22 @@ describe('Browser Content Tests', () => {
       `);
     });
 
-    httpServer = app.listen(0);
-    const address = httpServer.address() as AddressInfo;
-    baseUrl = `http://localhost:${address.port}`;
+    await new Promise((resolve, reject) => {
+      httpServer = app.listen(0, (error) => {
+        if (error) {
+          reject(error);
+          return;
+        }
+        const address = httpServer.address() as AddressInfo;
+        baseUrl = `http://localhost:${address.port}`;
+
+        resolve({});
+      });
+    });
   });
 
-  afterAll(async () => {
-    await httpServer.close();
+  afterAll(() => {
+    httpServer?.close();
   });
 
   beforeEach(async () => {
@@ -142,17 +152,11 @@ describe('Browser Content Tests', () => {
         arguments: { url: baseUrl },
       });
 
-      const result = await client.callTool({
-        name: 'browser_get_markdown',
-      });
-      expect(result.isError).toBe(false);
-      expect(result.content?.[0].text).toMatchInlineSnapshot(`
-        "Content Test Home
-
-        # Content Test Home
-
-         [Go to Page 1](/page1) [Go to Page 2](/page2)"
-      `);
+      await expect(
+        client.callTool({
+          name: 'browser_get_markdown',
+        }),
+      ).rejects.toThrowError(McpError);
     });
   });
 
@@ -180,14 +184,11 @@ describe('Browser Content Tests', () => {
         arguments: { url: baseUrl },
       });
 
-      const result = await client.callTool({
-        name: 'browser_get_text',
-      });
-      expect(result.isError).toBe(false);
-      expect(result.content?.[0].text).toMatchInlineSnapshot(`
-        "Content Test Home
-         Go to Page 1 Go to Page 2"
-      `);
+      await expect(
+        client.callTool({
+          name: 'browser_get_text',
+        }),
+      ).rejects.toThrowError(McpError);
     });
   });
 });