+++ packages/agent-infra/mcp-http-server/CHANGELOG.md
@@ -1,4 +1,16 @@
-# @agent-infra/mcp-shared
+# mcp-http-server
+
+## 1.2.4
+
+### Patch Changes
+
+- feat: tweak some features
+
+## 1.2.3
+
+### Patch Changes
+
+- chore: update mcp version
 
 ## 1.2.2