+++ packages/agent-infra/mcp-servers/commands/CHANGELOG.md
@@ -1,5 +1,27 @@
 ## 0.4.0 2024-12-10 - add logging
 
+## 1.2.14
+
+## 1.2.13
+
+## 1.2.12
+
+## 1.2.11
+
+### Patch Changes
+
+- feat: tweak some features
+- Updated dependencies
+  - mcp-http-server@1.2.4
+
+## 1.2.10
+
+### Patch Changes
+
+- chore: update mcp version
+- Updated dependencies
+  - mcp-http-server@1.2.3
+
 ## 1.2.9
 
 ## 1.2.8